% Auto-generated: do not edit by hand
\name{periodicTableInput}

\alias{periodicTableInput}

\title{PeriodicTableInput component}

\description{

}

\usage{
periodicTableInput(id=NULL, state=NULL, disabledElements=NULL, enabledElements=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{state}{Named list. A list of symbols that are currently highlighted as clicked}

\item{disabledElements}{Named list. A list of symbols to grey-out as disabled}

\item{enabledElements}{Named list. }
}
