% Auto-generated: do not edit by hand
\name{periodicTableInput}

\alias{periodicTableInput}

\title{PeriodicTableInput component}

\description{
MpPeriodicTableInput is a component that allows user to select some components from a periodic table
}

\usage{
periodicTableInput(id=NULL, state=NULL, disabledElements=NULL, enabledElements=NULL,
hiddenElements=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{state}{Named list. A list of symbols that are currently highlighted as clicked}

\item{disabledElements}{Named list. A list of symbols to grey-out as disabled}

\item{enabledElements}{Named list. A list of selected symbols}

\item{hiddenElements}{Named list. A list of hidden symbols}
}
